refactor: remove the dead preview module
`preview.py` had no production caller — the preview that ships is
`websocket.ws_preview_render`. Its 411 lines carried a stale 5th
widget-type table and a 4th WidgetState builder frozen at an old
widget/layout set, plus a third copy of the view→layout construction
loop, so it was drift waiting to be mistaken for the real seam.

The only live asset inside was the `MockHass` duck-type that
`test_websocket_preview` borrows; it moves to `tests/helpers.py`,
where a test double belongs. `scripts/mock_hass.py` keeps its own
sample-generation variant — that one auto-resolves `icon`
attributes, so sharing it would change what these tests observe.
